From 88de1af179ddac0167e9b31a0732d2a02c68efb8 Mon Sep 17 00:00:00 2001 From: Ben Gardner <gardner.ben@gmail.com> Date: Wed, 30 Mar 2016 15:09:26 -0500 Subject: [PATCH 20/32] Memtest86+: fix missing break in while statement
Coverity spotted a case of missing breaks in a switch.
The net effect is that instead of showing a CPU type guess of "586" or
"686", it shows "Unidentified Processor".

I'm not aware of any CPUs that would hit this code. This switch statement
may be effectively dead code, but we may as well have correct dead code.

Change-Id: I93349cf1ff9b9366610428c8e8ae3ae60d23b4a4
Signed-off-by: Ben Gardner <gardner.ben@gmail.com>
Found-by: Coverity (75579, 75580)
Reviewed-on: https://review.coreboot.org/14196
Tested-by: build bot (Jenkins)
Reviewed-by: Martin Roth <martinroth@google.com>