From 5cd1daf299348007c59f857d634b2e9aabd1c183 Mon Sep 17 00:00:00 2001 From: Ben Gardner <gardner.ben@gmail.com> Date: Wed, 30 Mar 2016 16:04:47 -0500 Subject: [PATCH 21/32] Memtest86+: clean up uninitialize and unused variable issues
Coverity spotted several cases where a variable might be used
uninitialized and a few where an assigned value was never used.

There were a few consecutive if() comparisons on the same variable that
should be if/else if chains.

There is also one instance where the code is clearly wrong, but without
the datasheet, it is unclear how to fix it. I marked that with a FIXME
comment.

Change-Id: I1a83770c1f365948c18d331814490afe43184a43
Signed-off-by: Ben Gardner <gardner.ben@gmail.com>
Found-by: Coverity (75575, 75588, 75589, 75590, 75591, 75592, 75594, 75595, 75596, 75597)
Reviewed-on: https://review.coreboot.org/14199
Tested-by: build bot (Jenkins)
Reviewed-by: Martin Roth <martinroth@google.com>